Add tolerance and circuit tripping.
It's super naive for now, and the module is starting to smell a bit, but
now the circuit will break after reaching a certain level of tolerance
for failure. It does not try to reset itself or allow test requests
through.

It does not respond to timeouts as an error. In fact, there is currently
no concept of a timeout.