Refactor changing state elements into an Agent as Breaker.Agent
The other, unchanging parts are simply returned as a circuit, a plain
map. This should probably become a struct.

We also implement a structure of buckets collected in a window with a
sum for calculating the health of the circuit. Additionally, we have
basic rolling for our rolling window scheme, but nothing is telling it
to roll.

The README needs to be updated to reflect the new rolling window scheme.