Wrap HTTP calls and recordings in a Task.
This makes the calls async by default, a more common use case for a
circuit breaker.

We also remove the use of HTTPotion.Base in Breaker. Since our needs
require us to implement some of their public API anyway, we weren't
gaining anything.

This is technically a breaking change, because we're changing the return
value and the interface to get the HTTP response, but we haven't even
made a release yet. :)